update to `xml-rpc` 0.2.0
this updates to the latest `xml-rpc` version. note that at the moment
this does not yet build since it hasn't been released yet.
this requires adnanademovic/xml-rpc-rs#13 to be merged & released first.

for the time being you can use this by using `[patch.crates-io]` in your
`Cargo.toml` and overwrite both `rosrust` and `xml-rpc`.

this update is needed to resolve various security vulnerabilities coming
from outdated versions of `hyper` which are being pulled in via
`xml-rpc`. with this, `cargo-audit` is happy again.